refactor: server
